Pull upstream change: move armor stand into own listener for MC 1.7 support
