+++ src/main/java/net/redstoneore/legacyfactions/listeners/AbstractConditionalListener.java
@@ -1,0 +1,9 @@
+package net.redstoneore.legacyfactions.listeners;
+
+import org.bukkit.event.Listener;
+
+public interface AbstractConditionalListener extends Listener {
+
+	boolean shouldEnable();
+	
+}
