+++ src/main/java/net/redstoneore/legacyfactions/listeners/FactionsBlockListener.java
@@ -3,7 +3,6 @@ package net.redstoneore.legacyfactions.listeners;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
@@ -11,7 +10,6 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.*;
-import org.bukkit.event.player.PlayerInteractAtEntityEvent;
 
 import net.redstoneore.legacyfactions.*;
 import net.redstoneore.legacyfactions.entity.Board;
@@ -61,21 +59,6 @@ public class FactionsBlockListener implements Listener {
 		if (PlayerMixin.canDoAction(event.getPlayer(), event.getBlock(), LandAction.DESTROY, false)) return;
 		
 		event.setCancelled(true);;
-	}
-	
-	// -------------------------------------------------- //
-	// ARMOR STAND PROTECTION
-	// -------------------------------------------------- //
-	
-	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-	public void playerInteractEntity(PlayerInteractAtEntityEvent event) {
-		Entity rightClicked = event.getRightClicked();
-		
-		if (rightClicked.getType() != EntityType.ARMOR_STAND) return;
-		
-		if (PlayerMixin.canDoAction(event.getPlayer(), rightClicked, LandAction.ENTITY, false)) return;
-		
-		event.setCancelled(true);
 	}
 	
 	// -------------------------------------------------- //
